Added page not found route, not authorized route, and some minor cosmetic tweaks
Going to any page that isn't in the code sends you to the PageNotFound.jsx.

Trying to access anything that you can't be authorized to sends you to NotAuthorized.jsx (tested by going to /not-authorized)

Added the pages to the App.jsx file.

Had to add the host property inside Vite's server config block in order access the web page in my browser (Was told of this fix in one of the classes)