Fix not authorized
Showing the NotAuthorized.jsx page should now function properly. Sawyer was able to work on the Cookies which enables the NotAuthorized functionality to work.

Now when you try to enter /main without proper credentials you are greeted with the NotAuthorized.jsx page. You do have to try and enter some credentials in.